make Proof general
