+++ storage/src/merkle/mod.rs
@@ -7,9 +7,11 @@
 
 pub mod location;
 pub mod position;
+pub mod proof;
 
 pub use location::{Location, LocationRangeExt};
 pub use position::Position;
+pub use proof::Proof;
 
 use core::fmt;
